+++ vue.config.js
@@ -15,6 +15,18 @@ module.exports = {
         from: path.join(__dirname, "./docs"),
         to: path.join(__dirname, "dist/docs"),
         toType: "dir"
+      },
+      {
+        context: 'node_modules/itk-vtk-viewer/dist/',
+        from: '**/*.png',
+        to: path.join(__dirname, "dist"),
+        toType: "dir",
+        force: true
+      },
+      {
+        from: path.join(__dirname, "node_modules/itk-vtk-viewer/dist/itk"),
+        to: path.join(__dirname, "dist/itk"),
+        toType: "dir"
       }
     ])]
   }